static-google-maps/script/requirements.txt
appdirs==1.4.4
attrs==22.1.0
Bottleneck==1.3.5
branca==0.6.0
brotlipy==0.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.0.4
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.0.5
cryptography==39.0.1
cycler==0.11.0
filelock==3.12.0
Fiona==1.9.1
folium==0.14.0
fonttools==4.25.0
fsspec==2023.5.0
GDAL==3.6.2
geopandas==0.12.2
huggingface-hub==0.14.1
idna==3.4
img2vec-pytorch==1.0.1
importlib-resources==5.2.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
mapclassify==2.5.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.2.1
munch==2.5.0
munkres==1.1.4
networkx==3.0
numexpr==2.8.4
numpy==1.24.3
packaging==23.1
pandas==1.5.3
Pillow==9.5.0
pip==23.0.1
pooch==1.4.0
pycparser==2.21
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyproj==2.6.1.post1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
Rtree==1.0.1
scikit-learn==1.2.2
scipy==1.10.1
setuptools==67.8.0
shapely==2.0.1
six==1.16.0
sympy==1.11.1
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tqdm==4.65.0
transformers==4.29.2
typing_extensions==4.6.2
urllib3==2.0.2
wheel==0.38.4
win-inet-pton==1.1.0
xyzservices==2022.9.0
zipp==3.11.0
